remote_travel_agents/ticket_agent/pyproject.toml
[project]
name = "ticket-agent"
version = "0.1.0"
description = "A2A Event Ticket Reservation Agent using CrewAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "dotenv>=0.9.9",
    "crewai>=0.118.0",
    "sse-starlette>=2.3.3",
    "jwcrypto>=1.5.6",
    "pyjwt>=2.10.1",
    "a2a-sdk>=0.2.16",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
